pyproject.toml
[project]
name = "langflowx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langflow>=1.6.9",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
